+++ CHANGELOG.md
@@ -17,6 +17,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Consolidated update modules to one module.
 - Update and install modules use logger instead of print statements for better logging and debugging.
 
+## [2.5.2-alpha] - 2026-04-30
+
+### Added
+
+- Added support for only `<hash>` in .sha512 checksum files (Fixes #297)
+
+### Fixed
+
+- Wrong hash type detection on .yml/.yaml checksum files
+
+
 ## [2.5.1-alpha] - 2026-04-30
 
 ### Added